Add files via upload
"Thrilled to share that I’ve designed a UI/UX prototype for a restaurant menu as part of my internship at CodSoft! 🍽️✨ This design aims to enhance the dining experience with an intuitive and visually appealing menu layout. Excited to see how it will elevate the restaurant's customer engagement!

LINK TO TAKS-2
https://www.figma.com/design/447kx1SpnSBCnnp8g7mC1y/Untitled?node-id=0-1&t=BQll3iQ3fmIAxgrM-1